integration: split pxeboot and dhclient tests
